requirements.txt
# Telegram Bot
python-telegram-bot>=21.0

# Scheduling reminders
APScheduler>=3.10.0

# Configuration & environment
python-dotenv>=1.0.0

# LLM (Groq)
groq>=0.4.0
